pyproject.toml
[tool.poetry]
name = "rest-api-client-framework"
version = "0.1.0-dev0"
description = "Generic Rest Api Client Framework"
authors = ["Quien Sabe <qs5779@mail.com>"]
license = "MIT"
readme = "README.md"
packages = [
    { include = "api_client"}
]

[tool.poetry.dependencies]
python = "^3.9"
urllib3 = "^2.2.2"
requests = "^2.32.3"
pydantic = { version = "^2.8.0", extras = [ "dotenv", "mypy" ] }
asyncio = "^3.4.3"

[tool.poetry.group.test]
optional = true

[tool.poetry.group.test.dependencies]
mypy = "^1.9"
wemake-python-styleguide = "^0.19"
flake8-pytest-style = ">=1.6,<3.0"
black = "^24.4.2"
isort = "^5.13.2"
pytest-cov = "^5.0.0"
coverage-badge = "^1.1.1"
types-requests = "^2.32.0.20240712"
pytest-httpserver = "^1.0.10"
pyfakefs = "^5.6.0"
pytest-docker = "^3.1.1"
doc8 = "^1.1.1"

[tool.poetry.group.docs]
optional = true

[tool.poetry.group.docs.dependencies]
sphinx = ">=5.2,<8.0"
sphinx-autodoc-typehints = ">=1.24,<3.0"
tomli = "^2.0"
m2r2 = "^0.3.3.post2"


[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
click = "^8.1.7"
yaml-settings-pydantic = {git = "https://github.com/acederberg/pydantic-settings-yaml.git", rev = "2.3.1"}
pdoc3 = "^0.11.1"
loguru = "^0.7.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
